requirements.txt
# main
Django==4.1.2
djangorestframework==3.14.0
requests==2.28.1

# debugging
ipython==8.5.0
